+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/common/callableId.kt
@@ -1,0 +1,9 @@
+package com.javiersc.kotlin.compiler.extensions.common
+
+import org.jetbrains.kotlin.name.CallableId
+import org.jetbrains.kotlin.name.FqName
+
+public fun String.toCallableId(): CallableId =
+    FqName(this).run { CallableId(parent(), shortName()) }
+
+public fun FqName.toCallableId(): CallableId = CallableId(parent(), shortName())