+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/fir/asFir.kt
@@ -1,0 +1,5 @@
+package com.javiersc.kotlin.compiler.extensions.fir
+
+import org.jetbrains.kotlin.fir.FirElement
+
+public inline fun <reified I : FirElement> FirElement?.asFir(): I? = this@asFir as? I?